+++ celescope/bulk_rna/__init__.py
@@ -1,4 +1,4 @@
-STEPS = ["sample", "starsolo", "cells", "split_fastq"]
+STEPS = ["sample", "starsolo", "cells", "split"]
 __ASSAY__ = "bulk_rna"
 REMOVE_FROM_MULTI = {
     "cells",